chore: move testing publisher to separate file
